Add license field to Forge project creator
