+++ src/main/resources/fileTemplates/j2ee/forge/mods.toml.ft
@@ -9,7 +9,7 @@ modLoader="javafml" #mandatory
 loaderVersion="[${FORGE_SPEC_VERSION},)" #mandatory This is typically bumped every Minecraft version by Forge. See our download page for lists of versions.
 # The license for you mod. This is mandatory metadata and allows for easier comprehension of your redistributive properties.
 # Review your options at https://choosealicense.com/. All rights reserved is the default copyright stance, and is thus the default here.
-license="All rights reserved"
+license="${LICENSE}"
 # A URL to refer people to when problems occur with this mod
 #issueTrackerURL="http://my.issue.tracker/" #optional
 # A list of mods - how many allowed here is determined by the individual mod loader
